done with email and change database

+++ db/models/user.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
       },
       phone: {
         allowNull: false,
-        type: DataTypes.INTEGER,
+        type: DataTypes.BIGINT,
       },
       img: {
         allowNull: false,
@@ -26,8 +26,8 @@ module.exports = (sequelize, DataTypes) => {
     },
     {
       sequelize,
-      modelName: "User",
-    }
+      modelName: 'User',
+    },
   );
   return User;
 };
